backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
SQLAlchemy>=2.0
asyncpg
# Google OAuth & Drive API
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
# Browser automation for cookie refresh
playwright
# Cloudflare R2 (S3-compatible)
boto3
